Time: 168 ms (50.99%), Space: 72.7 MB (68.83%) - LeetHub

+++ 3850-equal-sum-grid-partition-ii/3850-equal-sum-grid-partition-ii.java
@@ -11,18 +11,18 @@ class Solution {
         return true;
     }
     public boolean canPartitionGrid(int[][] grid) {
-        HashMap<Integer , List<int[]>> map = new HashMap<>();
+        HashMap<Long , List<int[]>> map = new HashMap<>();
         for(int i = 0; i < grid.length; i++){
             for(int j = 0; j < grid[0].length; j++){
-                map.put(grid[i][j] , new ArrayList<>());
+                map.put((long)grid[i][j] , new ArrayList<>());
             }
         }
         for(int i = 0; i < grid.length; i++){
             for(int j = 0; j < grid[0].length; j++){
-                map.get(grid[i][j]).add(new int[]{i , j});
+                map.get((long)grid[i][j]).add(new int[]{i , j});
             }
         }
-        int[][] ps = new int[grid.length][grid[0].length];
+        long[][] ps = new long[grid.length][grid[0].length];
         ps[0][0] = grid[0][0];
         for(int i = 1; i < grid.length; i++){
             ps[i][0] = ps[i-1][0]+grid[i][0];
@@ -35,14 +35,14 @@ class Solution {
                 ps[i][j] = ps[i-1][j]+ps[i][j-1]+grid[i][j]-ps[i-1][j-1];
             }
         }
-        int t = ps[ps.length-1][ps[0].length-1];
+        long t = ps[ps.length-1][ps[0].length-1];
         for(int i = 0; i < ps[0].length-1; i++){
-            int l = ps[ps.length-1][i];
-            int r = t-l;
+            long l = ps[ps.length-1][i];
+            long r = t-l;
             if(l == r)
             return true;
             else if(l < r){
-                int d = r-l;
+                long d = r-l;
                 if(map.containsKey(d)){
                     for(int[] a : map.get(d)){
                         if(a[1] <= ps[0].length-1 && a[1] > i && connected(a , 0 , ps.length-1 , i+1 , ps[0].length-1))
@@ -50,7 +50,7 @@ class Solution {
                     }
                 }
             }else{
-                int d = l-r;
+                long d = l-r;
                 if(map.containsKey(d)){
                     for(int[] a : map.get(d)){
                         if(a[1] >= 0 && a[1] <= i && connected(a , 0 , ps.length-1 , 0 , i))
@@ -60,12 +60,12 @@ class Solution {
             }
         }
         for(int i = 0; i < ps.length-1; i++){
-            int u = ps[i][ps[0].length-1];
-            int l = t-u;
+            long u = ps[i][ps[0].length-1];
+            long l = t-u;
             if(u == l)
             return true;
             else if(u < l){
-                int d = l-u;
+                long d = l-u;
                 if(map.containsKey(d)){
                     for(int[] a : map.get(d)){
                         if(a[0] > i && a[0] < ps.length && connected(a , i+1 , ps.length-1 , 0 , ps[0].length-1))
@@ -73,7 +73,7 @@ class Solution {
                     }
                 }
             }else{
-                int d = u-l;
+                long d = u-l;
                 if(map.containsKey(d)){
                     for(int[] a : map.get(d)){
                         if(a[0] >= 0 && a[0] <= i && connected(a , 0 , i , 0 , ps[0].length-1))
